refactor(web/mcp): split AuthActionBand into AdminAuthBand/WsAuthBand
Per AGENTS.md "Scope-isolated pages" rule: pages assemble scope-specific
modules rather than parameterizing one component with a role/scope prop.
AdminAuthBand binds admin org-grant APIs (including the
adminGetInstallEffective synthesized org-row state previously embedded in
MCPAdminDetailPanel) and forces callerRole='admin'/isOrgAdmin=true at
the call site. WsAuthBand binds workspace + per-user grant APIs and never
offers org-grant management — workspace admins must use the admin page,
matching spec §4. The shared AuthBandFrame stays scope-agnostic.
